Crate tokio_udp

UDP bindings for tokio.

Note: This crate is deprecated in tokio 0.2.x and has been moved intotokio::udp behind the udp feature flag.

This module contains the UDP networking types, similar to the standard library, which can be used to implement networking protocols.

The main struct for UDP is the UdpSocket, which represents a UDP socket. Reading and writing to it can be done using futures, which return the RecvDgram and SendDgram structs respectively.

For convenience it's also possible to convert raw datagrams into higher-level frames.

Structs

RecvDgram
A future used to receive a datagram from a UDP socket.

SendDgram
A future used to write the entire contents of some data to a UDP socket.

UdpFramed
A unified Stream and Sink interface to an underlying UdpSocket, using the Encoder and Decoder traits to encode and decode frames.

UdpSocket
An I/O object representing a UDP socket.